requirements.txt
livekit-agents>=0.12.9
livekit-plugins-silero>=0.7.4
livekit-plugins-azure>=0.5.3
livekit-plugins-openai>=0.10.16
livekit-plugins-google>=0.10.0
python-dotenv~=1.0
aiofile~=3.8.8
rich>=13.0.0
